Changes to the into the ending lighting etc
I changed the ending so you can fall through the floor. I changed the eniv a bit to be more clear. I added a sound which modulates when you get close. And I added a kill volume for one floor. And I updated the fonts so they change but I need to fix the other one for the game over screen.